refactor wallet: factor out base component
