+++ lib/wallet.js
@@ -1,25 +1,9 @@
 'use strict'
 
 const _ = require('lodash')
+const BaseComponent = require('./base-state')
 
-class Wallet {
-  constructor (state) {
-    this.state = state || []
-
-    this.sentSnap = false
-  }
-
-  update (d) {
-    const copy = JSON.parse(JSON.stringify(d))
-    const nSnap = this.parseSnap(copy)
-
-    const diff = this.diff(this.state, nSnap)
-
-    this.state = nSnap
-
-    return diff
-  }
-
+class Wallet extends BaseComponent {
   diff (o, n) {
     if (o.length === 0 && n.length === 0) return []
     if (o.length === 0 && n.length > 0) return n
@@ -33,10 +17,6 @@ class Wallet {
     })
 
     return updated
-  }
-
-  getState () {
-    return this.state
   }
 
   parseSnap (snap) {
